cleanup naming on mocks
