+++ consensus/src/authority/mocks/mod.rs
@@ -1,3 +1,3 @@
-pub mod actor;
-pub mod ingress;
+pub mod application;
 pub mod relay;
+pub mod supervisor;
